+++ tests/test_field.py
@@ -9,7 +9,7 @@ from configalchemy.types import Json, GenericConfigMixin
 
 class FieldTestCase(unittest.TestCase):
     def test_validate(self):
-        int_field = Field(name="TEST", default_value=0, value_type=int)
+        int_field = Field(name="TEST", default_value=0, annotation=int)
         for value in [b"0", "0"]:
             self.assertEqual(0, int_field.validate(value))
         with self.assertRaises(ValidateException) as e:
@@ -21,7 +21,7 @@ class FieldTestCase(unittest.TestCase):
 
     def test_union_type(self):
         optional_field = Field(
-            name="TEST", default_value=None, value_type=Optional[int]
+            name="TEST", default_value=None, annotation=Optional[int]
         )
         self.assertEqual(1, optional_field.validate(1))
         self.assertEqual(1, optional_field.validate("1"))
@@ -30,7 +30,7 @@ class FieldTestCase(unittest.TestCase):
         value_type = Json[list]
         default_value: value_type = []
         json_field = Field(
-            name="TEST", default_value=default_value, value_type=value_type
+            name="TEST", default_value=default_value, annotation=value_type
         )
         with patch.object(value_type, "__typecast__") as mock:
             self.assertEqual([1], json_field.validate([1]))
@@ -39,7 +39,7 @@ class FieldTestCase(unittest.TestCase):
 
         default_value: Json[List[int]] = [1, 2]
         json_field = Field(
-            name="TEST", default_value=default_value, value_type=Json[List[int]]
+            name="TEST", default_value=default_value, annotation=Json[List[int]]
         )
         self.assertEqual([1], json_field.validate(json.dumps([1])))
 
@@ -58,7 +58,7 @@ class FieldTestCase(unittest.TestCase):
 
         generic_config = TestGenericConfigMixin()
         generic_field = Field(
-            name="TEST", default_value=generic_config, value_type=None
+            name="TEST", default_value=generic_config, annotation=None
         )
         unittest_self.assertEqual(value, generic_field.validate(value))
         unittest_self.assertFalse(typecast.called)